Reimplementation of Alias * Since, alias args are directly handed over to variables::alias(), The values wrapped in Quotes are not stored correctly and can not be executed correctly
