+++ src/lib/builtins/variables.rs
@@ -1,138 +1,8 @@
 // TODO: Move into grammar
 
-use std::io::{self, Write};
 
 use shell::status::*;
 use shell::variables::Variables;
-use types::*;
-
-fn print_list(list: &VariableContext) {
-    let stdout = io::stdout();
-    let stdout = &mut stdout.lock();
-
-    for (key, value) in list {
-        let _ = stdout
-            .write(key.as_bytes())
-            .and_then(|_| stdout.write_all(b" = "))
-            .and_then(|_| stdout.write_all(value.as_bytes()))
-            .and_then(|_| stdout.write_all(b"\n"));
-    }
-}
-
-enum Binding {
-    InvalidKey(Identifier),
-    ListEntries,
-    KeyOnly(Identifier),
-    KeyValue(Identifier, Value),
-    Math(Identifier, Operator, f32),
-    MathInvalid(Value),
-}
-
-enum Operator {
-    Plus,
-    Minus,
-    Divide,
-    Multiply,
-}
-
-/// Parses alias as a `(key, value)` tuple.
-fn parse_alias(args: &str) -> Binding {
-    // Write all the arguments into a single `String`
-    let mut char_iter = args.chars();
-
-    // Find the key and advance the iterator until the equals operator is found.
-    let mut key = "".to_owned();
-    let mut found_key = false;
-    let mut operator = None;
-
-    // Scans through characters until the key is found, then continues to scan until
-    // the equals operator is found.
-    while let Some(character) = char_iter.next() {
-        match character {
-            ' ' if key.is_empty() => (),
-            ' ' => found_key = true,
-            '+' => {
-                if char_iter.next() == Some('=') {
-                    operator = Some(Operator::Plus);
-                    found_key = true;
-                }
-                break;
-            }
-            '-' => {
-                if char_iter.next() == Some('=') {
-                    operator = Some(Operator::Minus);
-                    found_key = true;
-                }
-                break;
-            }
-            '*' => {
-                if char_iter.next() == Some('=') {
-                    operator = Some(Operator::Multiply);
-                    found_key = true;
-                }
-                break;
-            }
-            '/' => {
-                if char_iter.next() == Some('=') {
-                    operator = Some(Operator::Divide);
-                    found_key = true;
-                }
-                break;
-            }
-            '=' => {
-                found_key = true;
-                break;
-            }
-            _ if !found_key => key.push(character),
-            _ => (),
-        }
-    }
-
-    let key: Identifier = key.into();
-
-    if !found_key && key.is_empty() {
-        Binding::ListEntries
-    } else {
-        let value: Value = char_iter.skip_while(|&x| x == ' ').collect();
-        if value.is_empty() {
-            Binding::KeyOnly(key)
-        } else if !Variables::is_valid_variable_name(&key) {
-            Binding::InvalidKey(key)
-        } else {
-            match operator {
-                Some(operator) => match value.parse::<f32>() {
-                    Ok(value) => Binding::Math(key, operator, value),
-                    Err(_) => Binding::MathInvalid(value),
-                },
-                None => Binding::KeyValue(key, value),
-            }
-        }
-    }
-}
-
-/// The `alias` command will define an alias for another command, and thus may be used as a
-/// command itself.
-pub(crate) fn alias(vars: &mut Variables, args: &str) -> i32 {
-    match parse_alias(args) {
-        Binding::InvalidKey(key) => {
-            eprintln!("ion: alias name, '{}', is invalid", key);
-            return FAILURE;
-        }
-        Binding::KeyValue(key, value) => {
-            vars.aliases.insert(key, value);
-        }
-        Binding::ListEntries => print_list(&vars.aliases),
-        Binding::KeyOnly(key) => {
-            eprintln!("ion: please provide value for alias '{}'", key);
-            return FAILURE;
-        }
-        _ => {
-            eprintln!("ion: invalid alias syntax");
-            return FAILURE;
-        }
-    }
-    SUCCESS
-}
 
 /// Dropping an alias will erase it from the shell.
 pub(crate) fn drop_alias<I: IntoIterator>(vars: &mut Variables, args: I) -> i32
@@ -204,6 +74,7 @@ mod test {
     use super::*;
     use parser::{expand_string, Expander};
     use shell::status::{FAILURE, SUCCESS};
+    use types::*;
 
     struct VariableExpander(pub Variables);
